:construction: [#2534] Move validation edit form utils
