+++ src/openforms/js/components/form/textfield.js
@@ -1,35 +1,10 @@
 import {Formio} from 'react-formio';
 
-import {
-  ADVANCED,
-  LOCATION,
-  PREFILL,
-  REGISTRATION,
-  TEXT_BASIC,
-  TEXT_VALIDATION,
-  TRANSLATIONS,
-  getValidationEditForm,
-} from './edit/tabs';
+import {DEFAULT_TEXT_TABS} from './edit/tabs';
 
 class TextField extends Formio.Components.components.textfield {
   static editForm() {
-    return {
-      components: [
-        {
-          type: 'tabs',
-          key: 'tabs',
-          components: [
-            TEXT_BASIC,
-            LOCATION,
-            ADVANCED,
-            getValidationEditForm(TEXT_VALIDATION),
-            REGISTRATION,
-            PREFILL,
-            TRANSLATIONS,
-          ],
-        },
-      ],
-    };
+    return {components: [DEFAULT_TEXT_TABS]};
   }
 }
